requirements.txt
--find-links https://download.pytorch.org/whl/cpu/torch_stable.html
torch
torchvision
opencv-python-headless
pillow
requests
supabase
numpy